+++ spec/support/repositories/active_record/computer_repository.rb
@@ -8,13 +8,13 @@ class ComputerRepositoryActiveRecord
   for_entity Computer
   default_provider :ActiveRecord
 
-  map_attribute :cpu,             'ActiveRecord::ComputerComponent::Hardware#cpu'
-  map_attribute :ram,             'ActiveRecord::ComputerComponent::Hardware#ram'
-  map_attribute :hdd,             'ActiveRecord::ComputerComponent::Hardware#hdd'
-  map_attribute :gfx,             'ActiveRecord::ComputerComponent::Hardware#gfx'
-  map_attribute :vendor,          'ActiveRecord::ComputerComponent::Hardware#vendor'
-  map_attribute :software_os,     'ActiveRecord::ComputerSoftware#os'
-  map_attribute :software_vendor, 'ActiveRecord::ComputerSoftware#vendor'
-  map_attribute :game_os,         'ActiveRecord::ComputerSoftware#os',     :via => :game_id
-  map_attribute :game_vendor,     'ActiveRecord::ComputerSoftware#vendor', :via => :game_id
+  map_attribute :cpu,             :to => 'ComputerComponent::Hardware#cpu'
+  map_attribute :ram,             :to => 'ComputerComponent::Hardware#ram'
+  map_attribute :hdd,             :to => 'ComputerComponent::Hardware#hdd'
+  map_attribute :gfx,             :to => 'ComputerComponent::Hardware#gfx'
+  map_attribute :vendor,          :to => 'ComputerComponent::Hardware#vendor'
+  map_attribute :software_os,     :to => 'ComputerSoftware#os'
+  map_attribute :software_vendor, :to => 'ComputerSoftware#vendor'
+  map_attribute :game_os,         :to => 'ComputerSoftware#os',     :via => :game_id
+  map_attribute :game_vendor,     :to => 'ComputerSoftware#vendor', :via => :game_id
 end