+++ spec/support/repositories/active_record/post_repository.rb
@@ -4,6 +4,6 @@ class PostRepositoryActiveRecord
   for_entity Post
   default_provider :ActiveRecord
 
-  map_attribute :author_name, 'ActiveRecord::Author#name', :via => :author_id
-  map_attribute :author_bio,  'ActiveRecord::Author#bio'
+  map_attribute :author_name, :to => 'Author#name', :via => :author_id
+  map_attribute :author_bio,  :to => 'Author#bio'
 end