+++ spec/support/repositories/active_record/user_repository.rb
@@ -4,8 +4,8 @@ class UserRepositoryActiveRecord
   for_entity User
   default_provider :ActiveRecord
 
-  map_attribute :last_name,      'ActiveRecord::User#surname'
-  map_attribute :driver_license, 'ActiveRecord::UserDriverLicense#number'
-  map_attribute :passport,       'ActiveRecord::UserPassport#number'
-  map_attribute :health_care,    'ActiveRecord::UserHealthCare#number'
+  map_attribute :last_name,      :to => 'User#surname'
+  map_attribute :driver_license, :to => 'UserDriverLicense#number'
+  map_attribute :passport,       :to => 'UserPassport#number'
+  map_attribute :health_care,    :to => 'UserHealthCare#number'
 end