+++ spec/support/repositories/callbacks_chaining_repository.rb
@@ -4,9 +4,9 @@ class CallbacksChainingRepository
   for_entity HeroUser
   default_provider :ActiveRecord
 
-  map_attribute :first_name, 'ActiveRecord::HeroUser#first_name'
-  map_attribute :last_name,  'Sequel::HeroUserLastName#last_name'
-  map_attribute :gender,     'Sequel::HeroUserLastName#gender'
+  map_attribute :first_name, :to => 'HeroUser#first_name'
+  map_attribute :last_name,  :to => 'HeroUserLastName#last_name', :provider => :Sequel
+  map_attribute :gender,     :to => 'HeroUserLastName#gender',    :provider => :Sequel
 
   before_save   :before_save_1
   before_save   :before_save_2
@@ -39,9 +39,9 @@ class CallbacksChainingPauseBeforeRepository
   for_entity HeroUser
   default_provider :ActiveRecord
 
-  map_attribute :first_name, 'ActiveRecord::HeroUser#first_name'
-  map_attribute :last_name,  'Sequel::HeroUserLastName#last_name'
-  map_attribute :gender,     'Sequel::HeroUserLastName#gender'
+  map_attribute :first_name, :to => 'HeroUser#first_name'
+  map_attribute :last_name,  :to => 'HeroUserLastName#last_name', :provider => :Sequel
+  map_attribute :gender,     :to => 'HeroUserLastName#gender',    :provider => :Sequel
 
   before_save   :before_save_1
   before_save   :before_save_2
@@ -74,9 +74,9 @@ class CallbacksChainingPauseAfterRepository
   for_entity HeroUser
   default_provider :ActiveRecord
 
-  map_attribute :first_name, 'ActiveRecord::HeroUser#first_name'
-  map_attribute :last_name,  'Sequel::HeroUserLastName#last_name'
-  map_attribute :gender,     'Sequel::HeroUserLastName#gender'
+  map_attribute :first_name, :to => 'HeroUser#first_name'
+  map_attribute :last_name,  :to => 'HeroUserLastName#last_name', :provider => :Sequel
+  map_attribute :gender,     :to => 'HeroUserLastName#gender',    :provider => :Sequel
 
   before_save   :before_save_1
   before_save   :before_save_2