+++ spec/support/repositories/sequel/user_repository.rb
@@ -4,8 +4,8 @@ class UserRepositorySequel
   for_entity User
   default_provider :Sequel
 
-  map_attribute :last_name,      'Sequel::User#surname'
-  map_attribute :driver_license, 'Sequel::UserDriverLicense#number'
-  map_attribute :passport,       'Sequel::UserPassport#number'
-  map_attribute :health_care,    'Sequel::UserHealthCare#number'
+  map_attribute :last_name,      :to => 'User#surname'
+  map_attribute :driver_license, :to => 'UserDriverLicense#number'
+  map_attribute :passport,       :to => 'UserPassport#number'
+  map_attribute :health_care,    :to => 'UserHealthCare#number'
 end